HackyVideoBullshitSite: Fixed stopping bug that was a bitch to debug because no warning was thrown about unregistered metatype, refactored and cleaned/simplified video segment folder management

+++ Solutions/hackyVideoBullshitSite/tools/FfmpegSegmentUploader/src/lib/ffmpegsegmentuploader.cpp
@@ -186,10 +186,13 @@ void FfmpegSegmentUploader::handleSegmentsEntryListFileModified()
 }
 void FfmpegSegmentUploader::killFfmpegProcessIfStillRunning()
 {
-    if(!m_FfmpegProcess->waitForFinished(1000))
+    if(m_FfmpegProcess->state() != QProcess::NotRunning)
     {
-        emit e("ffmpeg didn't finish within 30 seconds, so killing it");
-        m_FfmpegProcess->kill();
+        if(!m_FfmpegProcess->waitForFinished(1000))
+        {
+            emit e("ffmpeg didn't finish within 30 seconds, so killing it");
+            m_FfmpegProcess->kill();
+        }
     }
 }
 void FfmpegSegmentUploader::handleFfmpegProecssFinished(int exitCode, QProcess::ExitStatus exitStatus)
